+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - (Feature) (Platform) Sidecar Initial Resources
 - (Feature) (Platform) Extend Gateway Options
 - (Feature) (Platform) Sidecar Feature
+- (Feature) (Platform) Sidecar Auth
 
 ## [1.4.1](https://github.com/arangodb/kube-arangodb/tree/1.4.1) (2026-01-27)
 - (Bugfix) (Platform) Ensure Helm History limit in ArangoPlatformService